fix: repair merge regression breaking all SIMD sgemv backends + restore quality gates
The merge 1c55c79 resolved conflicts by keeping pre-clippy loop headers
with post-clippy loop bodies, leaving `*item` without a binding in the
AVX2, SSE2 and NEON sgemv_f32 implementations — the workspace no longer
compiled on x86_64 or aarch64. Same merge artifact in scirust-arena
(`_size:` vs `size` field, E0559).

- scirust-simd/dispatch.rs: restore iterator form of the three sgemv loops
- scirust-arena/slab.rs: use the actual `size` field name
- scirust-evo: take `&mut [f64]` instead of `&mut Vec<f64>`, iterator loops
- scirust-core: iterator loop in TT backward bias grad, use div_ceil()
- scirust-nas: drop usize-to-usize cast
- scirust-som tokenizer: add Default impl for StructuredTokenizer
- scirust-neuro-symbolic: targeted allow on Gaussian elimination loop
- cargo fmt over the workspace (58 pre-existing diffs in 14 files)

Verified: cargo check --workspace --all-targets, clippy --workspace --lib
-D warnings, cargo fmt --check all exit 0; cargo test --workspace passes
630 tests (0 failed, 18 ignored).

+++ scirust-nas/src/lib.rs
@@ -291,7 +291,7 @@ impl NasSearch {
                 // Mutate: change one random layer
                 if !child.layers.is_empty()
                 {
-                    let idx = ((gen * pop_size + i) as usize) % child.layers.len();
+                    let idx = (gen * pop_size + i) % child.layers.len();
                     let seed = (self.config.seed + (gen * pop_size + i) as u64) * 0x9E3779B9;
                     let rng = seed ^ (seed << 13);
                     let new_dim = self.config.min_hidden
